# visol/design/prototype/API.yaml
openapi: 3.0.0
info:
  title: ViSOL
  description: >-
    Visual Interactive Optimization Tool.
  contact: { }
  version: 1.0.0
servers:
  - url: 'https://visol1.paas.hosted-by-previder.com/api/v1'
    variables: { }
paths:
...
  '/changes/schedules/{vessel_id}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel whose Schedule should be/was changed.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Changes
      summary: Retrieve all Schedule's Changes
      operationId: retrieveAllScheduleChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChanges'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Changes
      summary: Delete a Schedule's Changes
      operationId: deleteScheduleChanges
      parameters:
        - name: from
          in: query
          description: Time to start deleting from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end deleting at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /changes/schedules:
    get:
      tags:
        - Changes
      summary: Retrieve all Schedules' Changes
      operationId: retrieveAllSchedulesChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChangesMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/changes/schedules/{vessel_id}/{date}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel which Schedule should be changed.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
      - name: date
        in: path
        description: >-
          Time at which the change was made. Can be formatted without colons and
          dashes.
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
    get:
      tags:
        - Changes
      summary: Retrieve a Schedule's Change
      operationId: retrieveScheduleChange
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChange'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Changes
      summary: Delete a Schedule's Change
      operationId: deleteScheduleChange
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
security:
  - bearerAuth: [ ]
components:
  schemas:
    Reason:
      type: string
      pattern: '^(?:.|\s)*$'
      maxLength: 1073741824
      nullable: true
      example: New vessel inbound (#142)
...
    Timestamp:
      type: string
      format: date-time
      pattern: >-
        ^(-?[0-9]+)-?(1[0-2]|0[1-9])-?(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):?([0-5][0-9]):?([0-5][0-9])Z$
      maxLength: 22
      example: 2022-05-11T16:32:35Z
    NullableSchedule:
      allOf:
        - $ref: '#/components/schemas/Schedule'
        - title: Nullable Schedule
          nullable: true
    Schedule:
      title: Schedule
      description: >-
        A Schedule object has the following attributes:


        + vessel: `1` (integer) - ID of the Vessel.
        
        + manual: `false` (boolean) - True if the Vessel should be planned at the supplied Berth and timestamp or false if it should be planned by the software.
        
        + berth: `20` (integer) - ID of the Berth the Vessel should be unloaded at.
        
        + start: `2022-05-12T07:23:00Z` (timestamp) - When the Vessel should start being unloaded.
        
        + end: `2022-05-13T21:41:00Z` (timestamp) - When the Vessel is expected to end being unloaded.
        
        
        The `berth` and `start` attributes are not necessary if the `type` is `automatic`.
        The `end` attribute will be calculated by the server.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        manual:
          type: boolean
        berth:
          $ref: '#/components/schemas/BerthID'
        start:
          $ref: '#/components/schemas/Timestamp'
        end:
          $ref: '#/components/schemas/Timestamp'
      additionalProperties: false
      required:
        - vessel
        - manual
        - berth
        - start
      example:
        vessel: 1
        manual: true
        berth: 30
        start: 2022-05-11T16:32:00Z
        end: 2022-05-11T21:26:00Z
...
    VesselID:
      title: Vessel ID
      description: >-
        An ID of a Vessel. Vessel IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    BerthID:
      title: Berth ID
      description: >-
        An ID of a Berth. Berth IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    ScheduleChange:
      title: Schedule Change
      description: >-
        A Vessel Change object has the following attributes:


        + vessel: `1` (integer) - The ID of the Schedule's Vessel. 

        + date: `2022-05-11T16:32:35Z` (timestamp) - The date at which the change was made.

        + old: `{"vessel":30,"manual":true,"berth":20,"start":"2022-05-15T16:32:00.000Z","end":"2022-05-15T21:26:00.000Z"}` (json) - The Schedule before being changed.

        + new: `{"vessel":30,"manual":false,"berth":20,"start":"2022-05-15T16:32:00.000Z","end":"2022-05-15T21:26:00.000Z"}` (json) - The Schedule after being changed.

        + reason: `Turn on automatic scheduling` (string) - The reason for the change.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        date:
          $ref: '#/components/schemas/Timestamp'
        old:
          $ref: '#/components/schemas/NullableSchedule'
        new:
          $ref: '#/components/schemas/Schedule'
        reason:
          $ref: '#/components/schemas/Reason'
      additionalProperties: false
      required:
        - vessel
        - date
        - old
        - new
        - reason
      example:
        vessel: 30
        date: 2022-05-11T16:32:35Z
        old:
          vessel: 30
          manual: true
          berth: 20
          start: 2022-05-15T16:32:00Z
          end: 2022-05-15T21:26:00Z
        new:
          vessel: 30
          manual: false
          berth: 20
          start: 2022-05-15T16:32:00Z
          end: 2022-05-15T21:26:00Z
        reason: Turn on automatic scheduling
    ScheduleChanges:
      title: Schedule Changes
      description: An array of Schedule Change objects.
      type: array
      items:
        $ref: '#/components/schemas/ScheduleChange'
      maxItems: 2147483647
      example:
        - vessel: 30
          date: 2022-05-11T16:32:35Z
          old:
            vessel: 30
            manual: true
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          new:
            vessel: 30
            manual: false
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          reason: Turn on automatic scheduling
        - vessel: 30
          date: 2022-05-11T21:26:24Z
          old:
            vessel: 30
            manual: false
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          new:
            vessel: 30
            manual: false
            berth: 18
            start: 2022-05-09T10:24:00Z
            end: 2022-05-10T16:17:00Z
          reason: Automatic scheduling job
    ScheduleChangesMap:
      title: Schedule Changes Map
      description: A map of Schedule Changes keyed by a Schedule's Vessel ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/ScheduleChanges'
      example:
        '30':
          - vessel: 30
            date: 2022-05-11T16:32:35Z
            old:
              vessel: 30
              manual: true
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            new:
              vessel: 30
              manual: false
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            reason: Turn on automatic scheduling
          - vessel: 30
            date: 2022-05-11T21:26:24Z
            old:
              vessel: 30
              manual: false
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            new:
              vessel: 30
              manual: false
              berth: 18
              start: 2022-05-09T10:24:00Z
              end: 2022-05-10T16:17:00Z
            reason: Automatic scheduling job
...
  responses:
    NoContent:
      description: 204 No Content
    NotFound:
      description: 404 Not Found
    Unauthorized:
      description: 401 Unauthorized
    Forbidden:
      description: 403 Forbidden
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: You must generate a JWT through the /token/obtain endpoint and then pass it as a Bearer token in the Authorization header.
...